Made 1 text component to house all my text components that are exported like an object
