--- src/components/SmallText.tsx
@@ -1,12 +1,0 @@
-interface Props {
-  children: string | number;
-  color?: string;
-}
-
-export default function SmallText({ children, color }: Props) {
-  return (
-    <h1 className="small" style={{ color }}>
-      {children}
-    </h1>
-  );
-}
