+++ src/pages/StoryTime/Users.tsx
@@ -1,10 +1,12 @@
-import Subtitle from "components/Subtitle";
+import Text from "components/Text";
 import Button from "components/Button";
 
 export default function Users() {
   return (
     <div className="leave">
-      <Subtitle subtitle="Users" optionalStyles={{ marginTop: "0.2em" }} />
+      <Text.Subtitle optionalStyles={{ marginTop: "0.2em" }}>
+        Users
+      </Text.Subtitle>
       <div className="users"></div>
       <Button text="Leave" />
     </div>
