+++ src/pages/StoryTime/index.tsx
@@ -1,14 +1,12 @@
-import Button from "components/Button";
-import Subtitle from "components/Subtitle";
-import Title from "components/Title";
 import React from "react";
 import Story from "./Story";
 import Users from "./Users";
+import Text from "components/Text";
 
 export default function () {
   return (
     <div className="container">
-      <Title />
+      <Text.Title />
       <div className="story-container">
         <Users />
         <Story />
